python/pyproject.toml
[tool.poetry]
name = "aoc"
version = "0.1.0"
description = ""
authors = ["Andreas Ellewsen <andreas@ellewsen.no>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
numpy = "^2.1.3"
python-dotenv = "^1.0.1"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
